extracted helper functions to utilities #7
